Add proto files to support the value this service will store
Summary:
In order to validate the signature properly, we must have a canonical way to
serialize map data. Therefore, we are switching to Protobufs from JSON as the
canonical way to serialize data.  While this requires additional dependencies,
protobufs was already needed for BIP70 support.  This is a good tradeoff to make
since JSON fields are unordered.

Test:
  go test ./...