Add BIP70 Pay-For-Put support
Summary:
The current state of this server is that you can put to keys which are the
Bitcoin(Cash) address that you control the private key for.  However, the result
of this alone will mean that nefarious parties will spam the server with PUTs for
various randomly generated keys in order to make running a server prohibitively
expensive.

Thus, this commit implements the start of a method for asking for BIP70 payment
upon the receipt of a PUT request.  The basic flow is:

1. Client attempts PUT /keys/<key> with an empty request body.
2. Server response with Status 402 (Payment Required) with a BIP70 payment details
  in the response body.  Payment details has the URL in the merchant data field.
3. Client generates a Payment and sends it to the pay_to field from the payment request.
4. Payment address validates the Payment and generates a Proof of Payment(PoP) Token.
  It responds with a HTTP 302 back to the original URL, a receipt token in the "Authorization"
  header, and a "Location" header with a "code" query-string amended to it with the PoP token.
5. Client follows the redirect, and PUTs the data it intended to put.

Test Plan:
go test ./...