Add support for configuration
Summary:
Currently, we have no way of specifying some basic configuration to the server.
This means we can't specify bind address, list of peers, or an HMAC secret. So,
in order to address that we add the use of the viper and cobra libraries for handling
command line arguments, and config files as yaml.

Closes #9

Test Plan:
Run by hand