Add, and enforce, TTL values for keys
Summary:
We want to ensure that we don't cache keys forever. Therefor, we
add a TTL value for each key.  If the TTL has expired, we will not
return the key. In future changes, we'll be looking up the value from
our peers in case they have a more recent version.

Test Plan:
```
go test ./...
```