Add kind field to the metdata rows.
Summary:
Based on discussions with the team, it seems that the kind field will
always be necessary. Therefore, it does not make sense to nest it under
the headers fields. Additionally, other headers likely are not useful.
However, they are left incase there is some unforeseen utility.

Test Plan:
```
go test ./...
```